consolidate terrain fns
